starting contact form

+++ admin/includes/view_all_posts.php
@@ -70,10 +70,20 @@
         <tbody>
         <?php
             if(isset($_GET['delete'])){
-                $the_post_id =$_GET['delete'];
-                $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
-                $delete_query = mysqli_query($connection, $query);
-                confirmQuery($delete_query);
+
+                if(isset($_SESSION['user_role'])){
+
+                    if($_SESSION['user_role'] == 'admin'){
+
+                        $the_post_id =$_GET['delete'];
+                        $query = "DELETE FROM posts WHERE post_id = {$the_post_id} ";
+                        $delete_query = mysqli_query($connection, $query);
+                        confirmQuery($delete_query);
+                    }
+
+                }
+
+             
             }
         ?>
